backend/requirements-voice.txt
edge-tts
flask
flask-cors
openai-whisper
ffmpeg-python
numba
scipy
